Revert "UIKit Chapter 02: Collection views and navigation"
