+++ 02-UIKit/README.md
@@ -2,6 +2,3 @@
 
 ## Chapter 1: UIKit essentials
  Set up your Xcode project, and compose your first collection view. Learn about collection views as you connect a data source.
-
-## Chapter 2: Collection views and navigation
- Create a new collection view to display more details about reminders. Create an accessible button that responds to user interaction. Navigate between your reminder list and detail views.
