--- 02-UIKit/Today/Today/Assets.xcassets/TodayListCellDoneButtonTint.colorset/Contents.json
@@ -1,52 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "extended-gray",
-        "components" : {
-          "alpha" : "0.500",
-          "white" : "0.000"
-        }
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "light"
-        }
-      ],
-      "color" : {
-        "color-space" : "display-p3",
-        "components" : {
-          "alpha" : "0.653",
-          "blue" : "0.000",
-          "green" : "0.000",
-          "red" : "0.000"
-        }
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "extended-gray",
-        "components" : {
-          "alpha" : "0.500",
-          "white" : "1.000"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
