--- 02-UIKit/Today/Today/Assets.xcassets/TodayPrimaryTint.colorset/Contents.json
@@ -1,41 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "platform" : "ios",
-        "reference" : "systemTealColor"
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "light"
-        }
-      ],
-      "color" : {
-        "platform" : "ios",
-        "reference" : "systemBlueColor"
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "platform" : "ios",
-        "reference" : "systemTealColor"
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
