--- 02-UIKit/Today/Today/Assets.xcassets/TodayProgressUpperBackground.colorset/Contents.json
@@ -1,54 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "extended-srgb",
-        "components" : {
-          "alpha" : "0.250",
-          "blue" : "0.839",
-          "green" : "0.820",
-          "red" : "0.820"
-        }
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "light"
-        }
-      ],
-      "color" : {
-        "color-space" : "extended-gray",
-        "components" : {
-          "alpha" : "0.250",
-          "white" : "1.000"
-        }
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "extended-srgb",
-        "components" : {
-          "alpha" : "0.250",
-          "blue" : "0.235",
-          "green" : "0.227",
-          "red" : "0.227"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
