+++ 02-UIKit/Today/Today/ReminderListViewController.swift
@@ -8,6 +8,9 @@
 import UIKit
 
 class ReminderListViewController: UICollectionViewController {
+    typealias DataSource = UICollectionViewDiffableDataSource<Int, String>
+    typealias Snapshot = NSDiffableDataSourceSnapshot<Int, String>
+    
     var dataSource: DataSource!
 
     override func viewDidLoad() {
@@ -16,7 +19,14 @@ class ReminderListViewController: UICollectionViewController {
         let listLayout = listLayout()
         collectionView.collectionViewLayout = listLayout
         
-        let cellRegistration = UICollectionView.CellRegistration(handler: cellRegistrationHandler)
+        let cellRegistration = UICollectionView.CellRegistration { (cell: UICollectionViewListCell,
+                                                                    indexPath: IndexPath,
+                                                                    itemIdentifier: String) in
+            let reminder = Reminder.sampleData[indexPath.item]
+            var contentConfiguration = cell.defaultContentConfiguration()
+            contentConfiguration.text = reminder.title
+            cell.contentConfiguration = contentConfiguration
+        }
         
         dataSource = DataSource(collectionView: collectionView) { (collectionView: UICollectionView,
                                                                    indexPath: IndexPath,
